Improve and add comments
Remove redundant UIElement.NewLine() method.
Fix hovered state for selected piece's square not being
cleared when canceling destination selection in UIFelliBoard.Play().

+++ FelliGame/UI/Entities/UIFelliSquare.cs
@@ -83,7 +83,7 @@ namespace FelliGame.UI.Entities
         {
             if (HasPiece)
             {
-                SetCursorDelta((int)(squareSize / 2), padding);
+                SetCursorDelta((squareSize / 2), padding);
 
                 Console.ForegroundColor = GetPieceConsoleColor(Square.Piece.Color);
